Refactored ConfigObject to improve flexibility and performance.
Using an adapter, the config information is passed to the adapter.
initialize() is called to parse, validate, and set the config params.
When ready to utilize the config param's, getParams() is called as needed.
The intent is to provide a better workflow and reduce performance the
performance hit by calling setParams() every time you need to get the
params.